add_websocket

+++ catch_at/backend/controllers/Favorites.js
@@ -4,32 +4,54 @@ const {
 
 module.exports.addToFavorites = async function (req, res) {
     console.log(req.body, " = body")
-    try {
-        const addToFavorites = await Favorite.create({
+    const isExist = await Favorite.findOne({
+        where: {
             subscriberId: req.body.subscriberId,
-            subscribedId: req.body.subscribedId,
-        })
-        return res.status(200).json(addToFavorites)
-    } catch (err) {
-        console.log("Promise Rejected", err);
+            subscribedId: req.body.subscribedId
+        }
+    })
+
+    if(!isExist){
+        try {
+            const addToFavorites = await Favorite.create({
+                subscriberId: req.body.subscriberId,
+                subscribedId: req.body.subscribedId,
+            })
+            return res.status(200).json(addToFavorites)
+        } catch (err) {
+            console.log("Promise Rejected", err);
+        }
     }
+    return res.status(304);
+
+    
 
 }
 
 
 module.exports.removeFromFavorites = async function (req, res) {
     console.log(req.body, " = body")
-    try {
-        const removeFromFavorites = await Favorite.destroy({
-            where: {
-                subscriberId: req.body.subscriberId,
-                subscribedId: req.body.subscribedId,
-            }
-        })
-        return res.status(200).json("Deleted Succesfully")
-    } catch (err) {
-        console.log("Promise Rejected", err);
+    const isExist = await Favorite.findOne({
+        where: {
+            subscriberId: req.body.subscriberId,
+            subscribedId: req.body.subscribedId
+        }
+    })
+    if(isExist){
+        try {
+            const removeFromFavorites = await Favorite.destroy({
+                where: {
+                    subscriberId: req.body.subscriberId,
+                    subscribedId: req.body.subscribedId,
+                }
+            })
+            return res.status(200).json("Deleted Succesfully")
+        } catch (err) {
+            console.log("Promise Rejected", err);
+        }
     }
+    return res.status(301)
+    
 
 }
 
@@ -37,6 +59,8 @@ module.exports.removeFromFavorites = async function (req, res) {
 module.exports.getAllFavorites = async function (req, res) {
     console.log("You are in getAllFavorites")
     try {
+        // const getAllFavorites = await User.findAll({attributes: ['id', 'username'], include: [{model: Favorite, attributes: ["subscriberId"]}, "id"]})
+        // return res.status(200).json(getAllFavorites)
         const getAllFavoritesOfUser = await Favorite.findAll({
             where: {
                 subscriberId: 1
